Module 10: Move get items to mongo
